agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "asyncio-mqtt==0.16.1",
    "grpcio==1.64.1",
    "grpcio-tools==1.64.1",
    "httpx==0.27.0",
    "protobuf==5.27.2",
    "psutil==5.9.8",
    "pyjwt[crypto]>=2.8.0",
    "pynvml==11.5.0",
]
